LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-006-expected.txt


PASS #target1 is scroll-adjusted in x axis only
PASS #target2 is scroll-adjusted in y axis only
PASS #target3 is scroll-adjusted in neither axis
